refactor(adapter-node): extract shared HTTP transport helpers to http-transport.ts
Move the adapter-agnostic retry/backoff/error-parsing helpers
(parseRetryAfter, retryAfterCause, retry, and the XML error-detail
helper, renamed s3ErrorDetail -> xmlErrorDetail) out of s3-http.ts
into a new http-transport.ts, so GcsHttpStorage can share them
instead of duplicating the retry loop and XML error parsing.
Behavior-preserving; s3-http.test.ts stays green unmodified.

The extraction moves PERMANENT_ERROR_CODES and parseRetryAfter out of
s3-http.ts, so repoint the four cross-reference comments still pointing
at their old home: the three credential-provider comments and the
constants.ts @see (also correcting its pre-existing protocol/src/storage
path, which never existed post adapter split). Stale pointers send
readers confidently to the wrong file, and verify:docs only validates
Markdown so nothing would flag them.
